refactored email storage tests
